Phase 4a: extract shared page-module render pipeline
The five page modules (search, shoppingCart, enrolledClasses, classDetail,
genericInstructor) each carried a ~95% identical two-phase renderPage (skeleton
mount + fetch/resolve/update). Extract it into
lib/content/shared/renderPipeline.ts (runRenderPipeline + reformatInitialLast);
each page module now declares only its selector, name/course extraction, mount
target, and Phase-1 class. Behavior preserved (the pipeline also dedupes matched
panels, which the comma-selector pages benefit from). ~440 lines removed.

NEEDS LIVE VERIFICATION on a real MyUCSC page (touches core content rendering).
