+++ src/lib/content/pages/classDetail.ts
@@ -1,26 +1,6 @@
-import {
-  getUIDFromJson,
-  fetchProfessorData,
-  fetchLocalResearchData,
-  fetchLocalClassesData,
-} from '@/lib/content/shared/professorResolver';
-import {
-  createMountPoint,
-  renderComponent,
-  unmountComponent,
-  isPlaceholderName,
-} from '@/lib/content/shared/mountHelper';
-import { getFirst } from '@/lib/format';
-import RatingBar from '@/components/RatingBar';
-import type {
-  PageConfig,
-  ProfessorData,
-  ProfessorBundle,
-  FetchProfessorDataResponse,
-  CampusProfile,
-  RmpTeacherNode,
-  RmpReview,
-} from '@/types';
+import { isPlaceholderName } from '@/lib/content/shared/mountHelper';
+import { runRenderPipeline } from '@/lib/content/shared/renderPipeline';
+import type { PageConfig } from '@/types';
 
 export const PAGE_CONFIG: PageConfig = {
   // `.PSGROUPBOXWBO` alone is too generic and makes the observer fire on nearly
@@ -100,110 +80,11 @@ export function getMountTarget(panel: Element): Element {
   );
 }
 
-/**
- * Full render pipeline for class detail pages.
- * Phase 1: Immediately render loading skeletons for all panels.
- * Phase 2: Fetch data and update with actual ratings.
- */
-export async function renderPage(): Promise<void> {
-  const panels = document.querySelectorAll(PAGE_CONFIG.panelSelector);
-  if (!panels.length) return;
-
-  // Phase 1: Immediately render loading skeletons
-  const mounts: {
-    mount: HTMLSpanElement;
-    name: string;
-    panel: Element;
-    course: string | null;
-  }[] = [];
-  for (const panel of panels) {
-    if (panel.classList.contains(PAGE_CONFIG.processedClass)) continue;
-    if (panel.querySelector('.rms-rating-bar-root')) continue;
-
-    const name = extractProfName(panel);
-    if (!name) continue;
-
-    const course = extractCourseCode(panel);
-    const target = getMountTarget(panel);
-    // Mark the panel processed regardless of fetch outcome, so a not-found
-    // professor (whose rating bar is removed below) is never reprocessed.
-    panel.classList.add(PAGE_CONFIG.processedClass);
-    const mount = createMountPoint(target, 'rms-rating-bar-root');
-    renderComponent(mount, RatingBar, { professorData: null, loading: true });
-    mounts.push({ mount, name, panel, course });
-  }
-
-  if (!mounts.length) return;
-
-  // Phase 2: Fetch data and update
-  const [researchTopics, classesTaught] = await Promise.all([
-    fetchLocalResearchData(),
-    fetchLocalClassesData(),
-  ]);
-
-  await Promise.allSettled(
-    mounts.map(async ({ mount, name, course }) => {
-      const uID = await getUIDFromJson(name);
-
-      let profileDict: FetchProfessorDataResponse | null = null;
-      try {
-        profileDict = await fetchProfessorData(uID, name);
-      } catch (error) {
-        // silently continue — error is non-critical
-      }
-
-      // Drop the { error } variant; only a successful bundle carries data.
-      const bundle: ProfessorBundle | null =
-        profileDict && !('error' in profileDict) ? profileDict : null;
-      // Some campus payloads nest a { success, data } wrapper under `data`; when
-      // the lookup failed, discard it. (Loose shape predating the typed bundle;
-      // cast is reconciled in a later phase.)
-      if (
-        bundle &&
-        (bundle.data as { success?: unknown } | null)?.success === false
-      ) {
-        bundle.data = null;
-      }
-
-      let profData: CampusProfile | null = null;
-      let rateMyProfessorData: RmpTeacherNode | null = null;
-      let reviews: RmpReview[] = [];
-      let researchTopicText: string | null = null;
-      let classesTaughtList: string[] | null = null;
-      let fullName: string | null = null;
-
-      if (bundle) {
-        profData = bundle.data;
-        rateMyProfessorData = bundle.rateMyProfessor;
-        reviews = bundle.reviews || [];
-        fullName = getFirst(profData?.cn);
-        if (fullName) {
-          researchTopicText = researchTopics[fullName];
-          classesTaughtList = classesTaught[fullName];
-        }
-      }
-
-      if (
-        !profData &&
-        !rateMyProfessorData &&
-        !researchTopicText &&
-        !classesTaughtList
-      ) {
-        unmountComponent(mount);
-        mount.remove();
-        return;
-      }
-
-      const professorData: ProfessorData = {
-        apiData: profData,
-        rateMyProfessor: rateMyProfessorData,
-        reviews,
-        localResearchTopic: researchTopicText,
-        localClassesTaught: classesTaughtList,
-        instructorName: name,
-        course,
-      };
-      renderComponent(mount, RatingBar, { professorData, loading: false });
-    })
-  );
+export function renderPage(): Promise<void> {
+  return runRenderPipeline({
+    config: PAGE_CONFIG,
+    extractProfName,
+    getMountTarget,
+    extractCourseCode,
+  });
 }
